Add files via upload
An extensive data analysis of the databases was performed and then we tried fitting the Machine learning models in the training dataset (80% of the data) and the one having the least error was fitted in the test data set which was the remaining data. The regularization parameter would be then used to minimize the error in the test data set and finally we would obtain the future value of the data.